refactor(auction-house): add execute_sale mod; refactor receipts
