+++ auction-house/program/src/receipt/mod.rs
@@ -1,10 +1,10 @@
 //! Create PDAs to to track the status and results of various Auction House actions.
 use crate::{
     constants::*,
+    errors::AuctionHouseError,
     id,
     instruction::{Buy, ExecuteSale, Sell},
     utils::*,
-    ErrorCode,
 };
 use anchor_lang::{prelude::*, AnchorDeserialize, AnchorSerialize};
 use solana_program::{sysvar, sysvar::instructions::get_instruction_relative};
@@ -102,14 +102,12 @@ pub struct PurchaseReceipt {
 #[derive(Accounts)]
 #[instruction(receipt_bump: u8)]
 pub struct PrintListingReceipt<'info> {
-    /// CHECK: Verified through CPI
     #[account(mut)]
     pub receipt: UncheckedAccount<'info>,
     #[account(mut)]
     pub bookkeeper: Signer<'info>,
     pub system_program: Program<'info, System>,
     pub rent: Sysvar<'info, Rent>,
-    /// CHECK: Verified through instruction ID
     #[account(address = sysvar::instructions::id())]
     pub instruction: UncheckedAccount<'info>,
 }
@@ -140,13 +138,10 @@ pub fn print_listing_receipt<'info>(
     let seller_trade_state = &prev_instruction_accounts[6];
     let metadata = &prev_instruction_accounts[2];
 
+    assert_program_listing_instruction(&prev_instruction.data[..8])?;
+
     let mut buffer = &prev_instruction.data[8..];
     let sell_data = Sell::deserialize(&mut buffer)?;
-
-    assert_program_instruction_equal(
-        &prev_instruction.data[..8],
-        [51, 230, 133, 164, 1, 127, 131, 173],
-    )?;
 
     assert_keys_equal(prev_instruction.program_id, id())?;
 
@@ -203,11 +198,9 @@ pub fn print_listing_receipt<'info>(
 /// Accounts for the [`cancel_listing_receipt` handler](fn.cancel_listing_receipt.html).
 #[derive(Accounts)]
 pub struct CancelListingReceipt<'info> {
-    /// CHECK: Verified through CPI
     #[account(mut)]
     pub receipt: UncheckedAccount<'info>,
     pub system_program: Program<'info, System>,
-    /// CHECK: Verified through instruction ID
     #[account(address = sysvar::instructions::id())]
     pub instruction: UncheckedAccount<'info>,
 }
@@ -227,13 +220,10 @@ pub fn cancel_listing_receipt<'info>(
 
     let trade_state = &prev_instruction_accounts[6];
 
-    assert_program_instruction_equal(
-        &prev_instruction.data[..8],
-        [232, 219, 223, 41, 219, 236, 220, 190],
-    )?;
+    assert_program_cancel_instruction(&prev_instruction.data[..8])?;
 
     if receipt_info.data_is_empty() {
-        return Err(ErrorCode::ReceiptIsEmpty.into());
+        return Err(AuctionHouseError::ReceiptIsEmpty.into());
     }
 
     assert_derivation(
@@ -258,14 +248,12 @@ pub fn cancel_listing_receipt<'info>(
 #[derive(Accounts)]
 #[instruction(receipt_bump: u8)]
 pub struct PrintBidReceipt<'info> {
-    /// CHECK: Verified through CPI
     #[account(mut)]
     receipt: UncheckedAccount<'info>,
     #[account(mut)]
     bookkeeper: Signer<'info>,
     system_program: Program<'info, System>,
     rent: Sysvar<'info, Rent>,
-    /// CHECK: Verified through instruction ID
     #[account(address = sysvar::instructions::id())]
     instruction: UncheckedAccount<'info>,
 }
@@ -306,7 +294,9 @@ pub fn print_bid_receipt<'info>(
 
     let token_account = match bid_type {
         BidType::PrivateSale => Some(token_account.pubkey),
+        BidType::AuctionPrivateSale => Some(token_account.pubkey),
         BidType::PublicSale => None,
+        BidType::AuctionPublicSale => None,
     };
 
     assert_derivation(
@@ -365,11 +355,9 @@ pub fn print_bid_receipt<'info>(
 /// Accounts for the [`cancel_bid_receipt` handler](fn.cancel_bid_receipt.html).
 #[derive(Accounts)]
 pub struct CancelBidReceipt<'info> {
-    /// CHECK: Verified through CPI
     #[account(mut)]
     receipt: UncheckedAccount<'info>,
     system_program: Program<'info, System>,
-    /// CHECK: Verified through instruction ID
     #[account(address = sysvar::instructions::id())]
     instruction: UncheckedAccount<'info>,
 }
@@ -389,13 +377,10 @@ pub fn cancel_bid_receipt<'info>(
 
     let trade_state = &prev_instruction_accounts[6];
 
-    assert_program_instruction_equal(
-        &prev_instruction.data[..8],
-        [232, 219, 223, 41, 219, 236, 220, 190],
-    )?;
+    assert_program_cancel_instruction(&prev_instruction.data[..8])?;
 
     if receipt_info.data_is_empty() {
-        return Err(ErrorCode::ReceiptIsEmpty.into());
+        return Err(AuctionHouseError::ReceiptIsEmpty.into());
     }
 
     assert_derivation(
@@ -420,20 +405,16 @@ pub fn cancel_bid_receipt<'info>(
 #[derive(Accounts)]
 #[instruction(receipt_bump: u8)]
 pub struct PrintPurchaseReceipt<'info> {
-    /// CHECK: Verified through CPI
     #[account(mut)]
     purchase_receipt: UncheckedAccount<'info>,
-    /// CHECK: Verified through CPI
     #[account(mut)]
     listing_receipt: UncheckedAccount<'info>,
-    /// CHECK: Verified through CPI
     #[account(mut)]
     bid_receipt: UncheckedAccount<'info>,
     #[account(mut)]
     bookkeeper: Signer<'info>,
     system_program: Program<'info, System>,
     rent: Sysvar<'info, Rent>,
-    /// CHECK: Verified through instruction ID
     #[account(address = sysvar::instructions::id())]
     instruction: UncheckedAccount<'info>,
 }
@@ -463,10 +444,7 @@ pub fn print_purchase_receipt<'info>(
     let mut buffer = &prev_instruction.data[8..];
     let execute_sale_data = ExecuteSale::deserialize(&mut buffer)?;
 
-    assert_program_instruction_equal(
-        &prev_instruction.data[..8],
-        [37, 74, 217, 157, 79, 49, 35, 6],
-    )?;
+    assert_program_purchase_instruction(&prev_instruction.data[..8])?;
 
     assert_keys_equal(prev_instruction.program_id, id())?;
 
@@ -510,7 +488,7 @@ pub fn print_purchase_receipt<'info>(
     )?;
 
     if listing_receipt_info.data_is_empty() || bid_receipt_info.data_is_empty() {
-        return Err(ErrorCode::ReceiptIsEmpty.into());
+        return Err(AuctionHouseError::ReceiptIsEmpty.into());
     }
 
     if purchase_receipt_info.data_is_empty() {
